refactor(connect): move bitcoin method definitions to corresponding folder
