+++ packages/connect-common/src/types/api/bitcoin/signTransaction.ts
@@ -4,8 +4,8 @@
  * User is asked to confirm all transaction details on Trezor.
  */
 
-import type { Params, Response } from '../params';
-import type { SignTransaction, SignedTransaction } from './bitcoin';
+import type { SignTransaction, SignedTransaction } from './common';
+import type { Params, Response } from '../../params';
 
 export declare function signTransaction(
     params: Params<SignTransaction>,
