+++ packages/connect-common/src/types/api/verifyMessage.ts
@@ -6,6 +6,6 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
 import type { Params, Response } from '../params';
-import type { VerifyMessage } from './bitcoin';
+import type { VerifyMessage } from './bitcoin/common';
 
 export declare function verifyMessage(params: Params<VerifyMessage>): Response<PROTO.Success>;
